Update _grib2io.py
First pass implementation of mapping grid data values to an enumerated key/code table.

Grib2Message.data() now has a new kwarg, map_keys=False/True, that signals the method to decode data values a string that the data value represents.

At this time, this only works for the scenario in which the key data is stored in the GRIB2 section2, local use section.  This is true for NWS Gridded MOS, NBM, and NDFD Wx grids.