Fix GDT bug in g2clib unpack3
Fixed bug in g2clib that was causing bad lat lon grid values.

The bug was in function unpack3 which is a wrapper for g2c function g2_unpack3.

In the grid definition template for regular and gaussian lat lon grids, a value is generally -1 and stored in a 4-byte (32-bit space) in the GRIB2 message.  The -1 value is packed as all 1s (signed 2's compliment).

However, in g2clib code space, starting with v1.6.4, all ints are 64-bit.  Therefore, a 32-bit int of -1 moved into 64-bit integer space yields at value of 4,294,967,295.

So we are now adding logic to check this for this value for regular and gaussian lat lon grids.