Update for g2clib Cython code
Added C preprocessing definitions to the docstring of the cdef extern
from g2c's grib2.h. Cython basically "appends" these definitions to the
actual grib2.h when building.

So now in the scope of grib2io and its g2clib Cython wrapper,
G2_PNG_ENABLED and G2_JPEG2000_ENABLED are always defined and set to 0
when not enabled or 1.