Descriptor classes for Section 5 metadata
Descriptor classes have been created for section 5 metadata and the data
representation templates along with a new function that returns a data
representation template class given the drtn.

Fixed some bugs present in the previous descriptor class work.

With this commit, the Grib2MessageBase.decode() method is no longer
needed and will be cleaned up or removed in the future.

Cleaned up and fixed isses in Grib2MessageBase.grid() relating to the
new descriptor classes.

This commit references #40.