BIG Update for _grib2io.py
Added durationa and validDate class variables to Grib2MessageBase.

Update open.select() method to support refDate, leadTime, and duration
as datetime objects.

On Grib2Message creation, only the metadata sections of the GRIB2
message are read from file.  The data section will be read only when the
data() method is called.

Clean up of decode() method.  It will be removed completely in a future
commit.

This commit references #1, #39, #40