More updates after Adam's PR
The commit message title does not imply that Adam's PR introduced bugs
or bad code, but simply to finish some of the ideas he implemented in
the PR.

Updated sections 4 and 5 descriptor classes to directly use data from
section4/5 array

Added validDate and descriptor classes for _fixedsfc*info private attrs.

*IMPORTANT* Added 'a' and 'b' keys to the projParameters dictionary.
These hold either the Earth's radius if a sphere or the major and minor
axes if Oblate Spheriod.