Update few minor changes for grib2io
Changing __repr__ for Grib2Message class to provide all attribute keys
and values. This is output is similar to degrib where each line is
prefixed with the GRIB2 section number, then attr name and value.

Updated functions to provide the 1 attr for section 6.

Added the template number attr name for sections 3, 4, and 5.