During install added LD_LIBRARY_PATH directories to find_library.
Now uses ctypes.util.find_library for Mac and Windows because it
returns the full path to the requested library.  Cross fingers
that works.  ctypes.util.find_library doesn't return the full
path for Linux and so adopted what was already written, along
with adding the search through the directories in
LD_LIBRARY_PATH.

Added an error message if libg2c couldn't be found.