Set kwargs["_xarray_backend"], an assert to ValueError, removed if.
If not using the xarray backend, in some situations the
kwargs["_xarray_backend"] throws a KeyError.  Fixed that if not
available setting to False, which is the default.

If opening a GRIB v1 file, there was an AssertionError which
is changed to raise an error and offer an error message.

Removed an unnecessary if block that tested for a header which
is now done in a loop immediately before.