MRMS GRIB2 File (#127)
* Adding MRMS GRIB2 tables.

This commit provides support for MRMS GRIB2 tables. MRMS uses the
GRIB2 discipline value of 209.  The included UserTable_MRMS_v12.2.csv
contains more information than what is needed, so in time, other
attributes from the table can be added.

This commit references issue #120

* Adding section4_discipline209.py

This commit adds section4_discipline209.py which contains the MRMS
variable parameter dictionaries.

This commit references issue #120

* Adding support for Gzip GRIB2 files.

This commit added support for reading GRIB2 files that have also been
gzipped.  Some datasets gzip GRIB2 files even though there is no benefit
in doing so.

When using the grib2io xarray backend, an error will be raised when
using Gzipped GIRB2 files.  This is a restiction within xarray.

This commit references issue #119

* Set kwargs["_xarray_backend"], an assert to ValueError, removed if. (#126)

If not using the xarray backend, in some situations the
kwargs["_xarray_backend"] throws a KeyError.  Fixed that if not
available setting to False, which is the default.

If opening a GRIB v1 file, there was an AssertionError which
is changed to raise an error and offer an error message.

Removed an unnecessary if block that tested for a header which
is now done in a loop immediately before.

Co-authored-by: Tim Cera <tim@cerazone.net>

---------

Co-authored-by: Eric Engle <EricEngle-NOAA@users.noreply.github.com>
Co-authored-by: TimothyCera-NOAA <154259143+TimothyCera-NOAA@users.noreply.github.com>
Co-authored-by: Tim Cera <tim@cerazone.net>