Add support for PDT 4.32 and change PDT class structure (#130)
* Add support for PDT 4.32 and change PDT class structure

This commit provides support for PDT 4.32.

This is the first PDT that deviates from the "base" structure of
PDT 0.  In 4.32, the fixed surface attributes are not present.
Therefore, need to remove section 4 attrs from the Grib2Message class
and define elsewhere.

The solution provided here is to create 2 new PDT-related template
classes as "base" classes.  These are ProductDefinitionTemplateBase and
ProductDefinitionTemplateSurface and they can be inherited by the actual
PDT classes.

Also here is a change to the wgrib2-formatted level string to provide
the string 'no_level'.

This commit references #128

* Update for attrs for PDT 4.32

Added satellite band attributes for when numberOfContributingSpectralBands is greater than 0.

* Add support for PDTN 4.31

PDTN 4.31 is similar to 4.32

Syntax clean up for class definitions

This commit references issue #128

---------

Co-authored-by: Eric Engle <EricEngle-NOAA@users.noreply.github.com>