132 properly mask output data when interpolating from smaller to larger domain (#133)
* Update for _grib2io.py

This commit was originally for 1) below, but in that work, a few
more bug were discovered and will be listed here.

1) When interpolating from a smaller domain to a larger domain grid. The
larger, output grid does not get properly filled with NaNs/missing values.
A TODO here is to be able to handle the various types of missing value
management.  In simple packing, a bitmap is used and for complex packing,
primary and secondary (optional) missing values are used.

2) Fixed issue in pack() where grids with NaNs were not being properly
cleaned up before packing in g2c, especially when a new message is being
packed (i.e. from interpolation). NOTE that g2c cannot handle NumPy
NaNs. This is WIP to handle more conditions.

3) Added drtn=None kwarg to Grib2Message.interpolate() method to allow
user control over how the interpolated data are packed. This is another
WIP to handle certain conditions. NOTE that is drtn=None, then the
packing scheme from the source message is used, but as a TODO, we need
to "sanitize" the DRT from the source grid.

This commit references issue #132

* Update _grib2io.py

Cleaned up to_bytes() method and added new Grib2Message method,
validate().

Also cleaned up trailing spaces.

This commit references issue #132

* Docstring updates.

[skip ci]

---------

Co-authored-by: Eric Engle <EricEngle-NOAA@users.noreply.github.com>