First commit to expose bitmap.
The GRIB2 message bitmap (if available) is now exposed as a public
attribute of the Grib2Message class with name, bitmap.

NOTE: The bitmap attribute will not be populated with bitmap data
until data are unpacked, otherwise it is None.

NOTE: When flush_data() method is called, the bitmap is flushed too.