Update _grib2io.py
When looking for the GRIB2 header string "GRIB", changed the
struct.unpack format from ">i" to ">I", an unsigned integer. The
integer representation of "GRIB" should never be negative.

This fixes cases where other data could be in the file and a given
4 bytes represented as ">i" could be negative which raises an error.
Instead, we just want to skip ahead and continue looking for "GRIB".