Update for Python 3.13 support
Big update for tables.get_varinfo_from_table(). Here we removed the
exec() which is a security risk and introduced a safer way to get
the var information dictionaries via importlib; new private function
to store tables in a dictionary so we do not have to reimport if
table already exists; no longer adding the tables to locals()
which was never ideal and not working in 3.13.

For Grib2Metadata class, made definition a property.

Updated GitHub workflows to test against 3.13.

Updated pyproject.toml to include 3.13 information.