Significant Update for xarray_backend.py
This commit allows for the backend to provide a common dimension name
for GRIB2 messages where more than 1 attribute would provide the dimensionality.

For example, probability messages (pdtn 5 or 9), these messages have the following
attrs used as dimensions: thresholdLowerLimit, thresholdUpperLimit, but these
are valid for 1 message, basically we are adding an extra dimension.

So here we introduce a "threshold" dimension (and coordinate), and thresholdLowerLimit
and thresholdUpperLimit serve as coordinates, dimensioned by "threshold".

Also, this commit adds the ability to add another node for Xarray DataTrees by the
typeOfProbability attribute, providing node name "prob_<typeOfProbability>".