Data shape and type checking when creating GRIB2 message (#214)
* Add shape and type checking when creating new Grib2Message

The setter property for Grib2Message.data now checks that the array
you are putting into the object is the same shape as the nx and ny
attributes for the message from section3.

The array will automatically be converted to float32 and ensure that
the memory layout is continguous.

* Update docs

[skip ci]

* Update Grib2Message data setter

Added further logic to check if attributes nx and ny of the
Grib2Message object are zero (likely a new message), then to use
the data shape to set these.

[skip ci]

* Update Grib2Message data setter

Updated tests and docs related to subject.

---------

Co-authored-by: Eric Engle <EricEngle-NOAA@users.noreply.github.com>