224 interp to stations out of grid domain (#226)
* Adding latlon_to_ij() in iplib Cython module

Added gdswzd() function definition in iplib.h

[skip ci]

* Add grib2io.utils.latlon_to_ij()

Added new function in utils which calls the iplib function.

Changed missing_value default to np.nan. This is OK even though np.nan
bit representation will be passed into Fortran.

[skip ci]

* Update _grib2io.py

When interpolating to stations, added logic to determine if the stations are
outside the grid domain and when this is true, to set the interpolated data
to np.nan.

The logic added will properly work for a single grid (ny, nx) or collection of
2D grids (i.e. 3D array (:, ny, nx) and works for scalar and vector interpolation.

* Added test for station coords.

* Update out of grid domain mask logic

Added test for interp to stations when a station is out of the grid domain.

---------

Co-authored-by: Eric Engle <EricEngle-NOAA@users.noreply.github.com>